+++ packages/ai/src/components/ai-service-module/ai-service-module-assistant-expanded.tsx
@@ -134,7 +134,7 @@ export const AIServiceModuleAssistantExpanded = React.forwardRef<HTMLDivElement,
         >
           <Icon icon={<X />} />
         </Button>
-        <InputSection sessionId={sessionId} variant="inset" autoFocus />
+        <InputSection sessionId={sessionId} variant="multiline" autoFocus />
       </div>
     );
   }