+++ packages/ai/src/components/input-section/input-section.tsx
@@ -1,4 +1,3 @@
-import { Input } from '@sk-web-gui/forms';
 import { cx } from '@sk-web-gui/utils';
 import React from 'react';
 import { useAssistantStore } from '../../assistant-store';
@@ -6,6 +5,9 @@ import { useChat } from '../../hooks';
 import { InputSectionButton } from './input-section-button';
 import { InputSectionInput, InputSectionInputProps } from './input-section-input';
 import { InputSectionWrapper } from './input-section-wrapper';
+import { ChatInput } from '../chat-input';
+import type { ChatInputProps } from '../chat-input/chat-input';
+import { handleMapToolbar } from '../chat-input/utils';
 
 export interface InputSectionDefaultProps {
   /**
@@ -13,22 +15,27 @@ export interface InputSectionDefaultProps {
    */
   shadow?: boolean;
   /**
-   * @default default
+   * @default multiline
    */
-  variant?: 'default' | 'inset';
+  variant?: 'multiline' | 'singleline';
   /**
    * @default false
    */
   isMobile?: boolean;
 }
+
 export interface InputSectionProps extends React.ComponentPropsWithoutRef<'form'>, InputSectionDefaultProps {
   sessionId?: string;
   onSendQuery?: (query: string) => void;
   placeholder?: InputSectionInputProps['placeholder'];
-  onChangeValue?: InputSectionInputProps['onChange'];
+  onChangeValue?: InputSectionInputProps['onChange'] | ChatInputProps['onChangeValue'];
   value?: string;
   button?: React.JSX.Element;
   autoFocus?: boolean;
+  /**
+   * Toolbar - only available with variant "multiline"
+   */
+  toolbar?: ChatInputProps['toolbar'];
 }
 
 export const InputSection = React.forwardRef<HTMLFormElement, InputSectionProps>((props, ref) => {
@@ -44,6 +51,7 @@ export const InputSection = React.forwardRef<HTMLFormElement, InputSectionProps>
     variant = 'default',
     button = <InputSectionButton isMobile={isMobile} variant={variant} />,
     autoFocus,
+    toolbar,
     ...rest
   } = props;
   const [query, setQuery] = React.useState<string>('');
@@ -52,21 +60,28 @@ export const InputSection = React.forwardRef<HTMLFormElement, InputSectionProps>
   const info = useAssistantStore((state) => state.info);
 
   const inputref = React.useRef<HTMLInputElement>(null);
+  const textarearef = React.useRef<HTMLTextAreaElement>(null);
 
   React.useEffect(() => {
-    if (autoFocus && inputref.current) {
-      inputref.current.focus();
+    if (autoFocus) {
+      inputref?.current?.focus?.();
+      textarearef?.current?.focus?.();
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [autoFocus, inputref.current]);
+  }, [autoFocus, inputref.current, textarearef.current]);
 
-  const handleOnChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  const handleOnChange = (event: React.ChangeEvent<HTMLInputElement> & React.ChangeEvent<HTMLTextAreaElement>) => {
     onChangeValue?.(event);
+
     setQuery(event.target.value);
   };
 
   const handleSubmit = (event: React.FormEvent) => {
     event.preventDefault();
+    submit();
+  };
+
+  const submit = () => {
     if (onSendQuery) {
       onSendQuery(query);
     } else {
@@ -77,17 +92,49 @@ export const InputSection = React.forwardRef<HTMLFormElement, InputSectionProps>
     setQuery('');
   };
 
+  const handleEnter = (event: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if (event.key === 'Enter' && !event.shiftKey && !event.ctrlKey && !event.altKey) {
+      event.preventDefault();
+      submit();
+    }
+  };
+
   return (
     <form ref={ref} className={cx('sk-ai-inputsection', className)} onSubmit={handleSubmit} {...rest}>
-      <InputSectionWrapper shadow={shadow} variant={variant}>
-        <InputSectionInput
-          placeholder={placeholder ?? `Skriv till ${info ? info.name : 'assistanten'}`}
-          onChange={handleOnChange}
-          value={value ?? query}
-          isMobile={isMobile}
-          ref={inputref}
-        />
-        {variant === 'inset' ? <Input.RightAddin>{button}</Input.RightAddin> : button}
+      <InputSectionWrapper shadow={shadow}>
+        {variant === 'singleline' ? (
+          <>
+            <InputSectionInput
+              placeholder={placeholder ?? `Skriv till ${info ? info.name : 'assistanten'}`}
+              onChange={handleOnChange}
+              value={value ?? query}
+              isMobile={isMobile}
+              ref={inputref}
+            />
+            {button}
+          </>
+        ) : (
+          <ChatInput.Wrapper>
+            <ChatInput.Textarea
+              placeholder={placeholder ?? `Skriv till ${info ? info.name : 'assistanten'}`}
+              onChange={handleOnChange}
+              value={value ?? query}
+              onKeyDown={handleEnter}
+            ></ChatInput.Textarea>
+            {toolbar ? (
+              <ChatInput.Toolbar>
+                {toolbar.map((group, index) => (
+                  <React.Fragment key={`toolbar-group-${index}`}>
+                    {handleMapToolbar(group, index, (toolbar?.length ?? 0) - 1)}
+                  </React.Fragment>
+                ))}
+              </ChatInput.Toolbar>
+            ) : (
+              <></>
+            )}
+            <ChatInput.Submitbutton type="submit" />
+          </ChatInput.Wrapper>
+        )}
       </InputSectionWrapper>
     </form>
   );