+++ packages/accordion/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.3](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@3.0.2...@sk-web-gui/accordion@3.0.3) (2026-02-24)
+
+**Note:** Version bump only for package @sk-web-gui/accordion
+
+## [3.0.2](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@3.0.1...@sk-web-gui/accordion@3.0.2) (2026-02-23)
+
+**Note:** Version bump only for package @sk-web-gui/accordion
+
+## [3.0.1](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@3.0.0...@sk-web-gui/accordion@3.0.1) (2026-02-02)
+
+### Bug Fixes
+
+- render issues ([f2d2137](https://github.com/Sundsvallskommun/web-shared-components/commit/f2d2137e5551dc69bc7462955754534a4d675ec2))
+
 # [3.0.0](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@2.1.9...@sk-web-gui/accordion@3.0.0) (2025-11-27)
 
 ### Features