+++ packages/cookie-consent/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.14](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/cookie-consent@2.1.13...@sk-web-gui/cookie-consent@2.1.14) (2026-02-24)
+
+**Note:** Version bump only for package @sk-web-gui/cookie-consent
+
+## [2.1.13](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/cookie-consent@2.1.12...@sk-web-gui/cookie-consent@2.1.13) (2026-02-23)
+
+**Note:** Version bump only for package @sk-web-gui/cookie-consent
+
+## [2.1.12](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/cookie-consent@2.1.11...@sk-web-gui/cookie-consent@2.1.12) (2026-02-02)
+
+**Note:** Version bump only for package @sk-web-gui/cookie-consent
+
 ## [2.1.11](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/cookie-consent@2.1.10...@sk-web-gui/cookie-consent@2.1.11) (2025-11-27)
 
 **Note:** Version bump only for package @sk-web-gui/cookie-consent