+++ packages/menu-vertical/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.10](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/menu-vertical@2.2.9...@sk-web-gui/menu-vertical@2.2.10) (2026-02-24)
+
+**Note:** Version bump only for package @sk-web-gui/menu-vertical
+
+## [2.2.9](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/menu-vertical@2.2.8...@sk-web-gui/menu-vertical@2.2.9) (2026-02-23)
+
+**Note:** Version bump only for package @sk-web-gui/menu-vertical
+
 ## [2.2.8](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/menu-vertical@2.2.7...@sk-web-gui/menu-vertical@2.2.8) (2025-09-15)
 
 **Note:** Version bump only for package @sk-web-gui/menu-vertical